relocate acceptance_specs within the test acceptance unix location
