+++ test/acceptance/unix/nodesets/centos-6-x64.yml
