+++ test/acceptance/unix/nodesets/debian-6-x64.yml
